SAGE-792: convert to shutdown service and add `--defer` option
Added `--defer` option signals to only set the hostname in the
`/etc/hostname` file and to NOT change the current running hostname.
Therefore, on system shutdown the service is executed to prepare
the system with the correct hostname on the following boot. This ensures
that all systems services (i.e. NetworkManager, etc.) have the correct
hostname on boot.